Ignore comments in test_fastq2fasta. New method.
The new method is based on mappy:
https://github.com/lh3/minimap2/blob/master/python/README.rst

It doesn't give access to the comment/description in fastx headers. For
that reason, I added a way to generate fasta without comments when
unwrapping and use this for the tests. This choice could be discussed,
and it would be easy to revert by setting `strip_comment=False` in
`Fastq2Fasta.unwrap_fasta`.