Added class to benchmark on multiple converters.
The goal is to be able to benchmark on various input files, for instance
of various sizes.
The times for the various converters can be weighted, for instance to
have the possibility to compensate for the short running times of
conversion of small files.
I'm not sure the weighting implementation is well working yet.