Added tests for gzipped input and in_gz decorator.
`@in_gz` marks a method with an `in_gz` attribute that can be used to
know if the method accepts gzipped input. `@compressor` implies
`@in_gz`.

I also added `@out_compressor`, that doesn't deal with the input.
`@out_compressor` is intended for methods that natively handle
compressed input, but doesn't imply `@in_gz`. So `@in_gz` needs to be
manually applied on top of `@out_compressor`. I suspect the `in_gz`
attribute will be lost if other decorators were applied on top of it.`